Refactor content_provider
