+++ examples/content_provider/content_provider/mod.rs
@@ -1,0 +1,19 @@
+mod imp;
+
+use gtk::{gdk, glib};
+
+glib::wrapper! {
+    pub struct ContentProvider(ObjectSubclass<imp::ContentProvider>) @extends gdk::ContentProvider;
+}
+
+impl ContentProvider {
+    pub fn new() -> Self {
+        glib::Object::new::<Self>(&[]).unwrap()
+    }
+}
+
+impl Default for ContentProvider {
+    fn default() -> Self {
+        Self::new()
+    }
+}
